+++ project_stage_state/models/__init__.py
@@ -2,4 +2,5 @@
 # Daniel Reis, 2014
 # GNU Affero General Public License <http://www.gnu.org/licenses/>
 
-from . import models
+from . import project_task
+from . import project_task_type
