+++ project_stage_state/models/project_task_type.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Daniel Reis, 2014
+# GNU Affero General Public License <http://www.gnu.org/licenses/>
+
+from odoo import models, fields
+
+
+class ProjectTask(models.Model):
+    _inherit = 'project.task'
+    state = fields.Selection(
+        related='stage_id.state', store=True, readonly=True)
